refactor(community): modularize frontend implementation
Refs #140

Refs #146

Community - Secure Community delete permissions with ownership-based UI and RLS #146 is only partially fixed and not fully checked. This commit adds frontend ownership-aware delete controls and guards, but database RLS still needs to be applied and verified.
